Moved modify-name-proposal to name commands and tested it locally.
