+++ x/name/client/cli/gov.go
@@ -25,6 +25,8 @@ const (
 	FlagTitle = "title"
 	// FlagDeposit is the flag for a deposit.
 	FlagDeposit string = "deposit"
+	// FlagMetadata is the flag for a deposit.
+	FlagMetadata string = "metadata"
 )
 
 // GetRootNameProposalCmd returns a command for registration with the gov module
@@ -107,92 +109,6 @@ $ %s tx gov submit-legacy-proposal \
 	}
 
 	cmd.Flags().String(FlagOwner, "", "The owner of the new name, optional (defaults to from address)")
-	cmd.Flags().BoolP(flagUnrestricted, "u", false, "Allow child name creation by everyone")
-	// proposal flags
-	cmd.Flags().String(FlagTitle, "", "Title of proposal")
-	cmd.Flags().String(FlagDescription, "", "Description of proposal")
-	cmd.Flags().String(FlagDeposit, "", "Deposit of proposal")
-	return cmd
-}
-
-func GetModifyNameProposalCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "modify-name-proposal [name] (--owner [address]) (--unrestrict) [flags]",
-		Short: "Submit a modify name creation governance proposal",
-		Long: strings.TrimSpace(
-			fmt.Sprintf(`Submit a modify name governance proposal along with an initial deposit.
-The proposal title and description must be provided through their respective flags.
-
-IMPORTANT: The restricted creation of sub-names will be enabled by default unless the unrestricted flag is included.
-The owner must approve all child name creation unless an alterate owner is provided.
-
-Example:
-$ %s tx gov submit-legacy-proposal \
-	modify-name-proposal \
-	<name> \
-	--unrestrict  \ 
-	--owner <address> \
-	--title "Proposal title" \
-	--description "Description of proposal" 
-	--from <key_or_address>
-			`,
-				version.AppName)),
-		Args: cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			proposalTitle, err := cmd.Flags().GetString(FlagTitle)
-			if err != nil {
-				return fmt.Errorf("proposal title: %w", err)
-			}
-			proposalDescr, err := cmd.Flags().GetString(FlagDescription)
-			if err != nil {
-				return fmt.Errorf("proposal description: %w", err)
-			}
-			proposalOwner, err := cmd.Flags().GetString(FlagOwner)
-			if err != nil {
-				return fmt.Errorf("proposal name owner: %w", err)
-			}
-			if len(proposalOwner) < 1 {
-				proposalOwner = clientCtx.GetFromAddress().String()
-			}
-			_, err = sdk.AccAddressFromBech32(proposalOwner)
-			if err != nil {
-				return err
-			}
-			depositArg, err := cmd.Flags().GetString(FlagDeposit)
-			if err != nil {
-				return err
-			}
-			deposit, err := sdk.ParseCoinsNormalized(depositArg)
-			if err != nil {
-				return err
-			}
-
-			content := types.ModifyNameProposal{
-				Title:       proposalTitle,
-				Description: proposalDescr,
-				Name:        strings.ToLower(args[0]),
-				Owner:       proposalOwner,
-				Restricted:  !viper.GetBool(flagUnrestricted),
-			}
-
-			msg, err := govtypesv1beta1.NewMsgSubmitProposal(&content, deposit, clientCtx.GetFromAddress())
-			if err != nil {
-				return err
-			}
-			if err = msg.ValidateBasic(); err != nil {
-				return err
-			}
-
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	cmd.Flags().String(FlagOwner, "", "The new owner of the name, optional (defaults to from address)")
 	cmd.Flags().BoolP(flagUnrestricted, "u", false, "Allow child name creation by everyone")
 	// proposal flags
 	cmd.Flags().String(FlagTitle, "", "Title of proposal")
